code/requirements.txt
openai==1.51.0sib_api_v3_sdk==7.6.0python-dotenv==1.0.1PyPDF2==3.0.1reportlab==4.2.5